refactor: remove empty folders, start docs improvements
