+++ elasticai/creator/nn/linear/layer.py
@@ -7,7 +7,7 @@ from elasticai.creator.base_modules.linear import Linear
 from elasticai.creator.base_modules.two_complement_fixed_point_config import (
     FixedPointConfig,
 )
-from elasticai.creator.vhdl.design_base.design import Design
+from elasticai.creator.vhdl.design.design import Design
 from elasticai.creator.vhdl.translatable import Translatable
 
 from .design import FPLinear as FPLinearDesign
