+++ elasticai/creator/nn/sequential/layer.py
@@ -3,7 +3,7 @@ from typing import cast
 
 import torch
 
-from elasticai.creator.vhdl.design_base.design import Design
+from elasticai.creator.vhdl.design.design import Design
 from elasticai.creator.vhdl.translatable import Translatable
 
 from .design import Sequential as _SequentialDesign
