+++ elasticai/creator/vhdl/design/signal.py
