+++ elasticai/creator/vhdl/shared_designs/rom/rom.tpl.vhd
